Decide to change all project structure to achieve more flexibility
